+++ src/main/java/lotto/domain/Lottos.java
@@ -2,19 +2,26 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.services.LottoNumberGenerator;
 
 public class Lottos {
     private List<Lotto> lottos;
     private final LottoNumberGenerator lottoNumberGenerator;
+    private Lotto winningNumbers;
 
     public Lottos(LottoNumberGenerator lottoNumberGenerator, int numberOfLottos) {
         this.lottoNumberGenerator = lottoNumberGenerator;
         this.lottos = create(numberOfLottos);
     }
 
-    public List<Lotto> getLottos(){
+    public List<Lotto> getLottos() {
         return lottos;
     }
+
+    public void setWinningNumbers(List<Integer> winningNumbers) {
+        this.winningNumbers = new Lotto(winningNumbers);
+    }
+
     private List<Lotto> create(int numberOfLottos) {
         List<Lotto> lottoBundle = new ArrayList<>();
         for (int count = 0; count < numberOfLottos; count++) {
@@ -23,6 +30,7 @@ public class Lottos {
         }
         return lottoBundle;
     }
+
     private Lotto generateNumber() {
         List<Integer> numbers = lottoNumberGenerator.generateNumbers();
         return new Lotto(numbers);
