+++ src/main/java/lotto/services/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.services;
 
 import java.util.List;
 import lotto.utils.NumberGenerator;
@@ -10,7 +10,7 @@ public class LottoNumberGenerator {
         this.numberGenerator = numberGenerator;
     }
 
-    public List<Integer> generateNumbers(){
+    public List<Integer> generateNumbers() {
         return numberGenerator.generate();
     }
 }
