+++ src/test/java/lotto/LottoNumberGeneratorTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import lotto.domain.Lotto;
-import lotto.domain.LottoNumberGenerator;
 import lotto.domain.Lottos;
+import lotto.services.LottoNumberGenerator;
 import lotto.utils.NumberGenerator;
 import lotto.utils.RandomNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
