+++ src/test/java/lotto/NumberComparatorTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.EnumMap;
 import java.util.List;
 import lotto.domain.Lotto;
-import lotto.domain.LottoNumberComparator;
 import lotto.enums.LottoPrize;
+import lotto.services.LottoNumberComparator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
